requirements.txt
# Core dependencies
numpy<1.20.0
torch>=1.3.0
mmcv>=1.3.17
mmdet>=2.14.0
mmsegmentation>=0.14.1

# MMDetection3D dependencies
lyft_dataset_sdk
networkx>=2.2,<2.3
numba==0.48.0
nuscenes-devkit
plyfile
scikit-image
tensorboard
trimesh>=2.35.39,<2.35.40

# Visualization and utilities
matplotlib
tqdm
PyYAML
opencv-python
pillow

# Optional dependencies
# Add any additional dependencies specific to your BEVFormer implementation 